requirements.txt
discord.py==2.7.1
python-dotenv==0.21.0